pyproject.toml
[project]
name = "petrosa-data-manager"
version = "1.0.0"
description = "Data integrity, intelligence, and distribution hub for the Petrosa trading ecosystem"
authors = [
    {name = "Petrosa Systems", email = "info@petrosa.com"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "nats-py>=2.7.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "motor>=3.3.0",
    "pymongo>=4.6.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "structlog>=24.1.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
    "opentelemetry-exporter-otlp>=1.22.0",
    "prometheus-client>=0.19.0",
    "httpx>=0.26.0",
    "python-json-logger>=2.0.7",
    "aiofiles>=23.2.1",
    "scipy>=1.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=24.1.0",
    "flake8>=7.0.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "bandit>=1.7.6",
    "types-aiofiles>=23.2.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_optional = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-v --cov=data_manager --cov-report=term --cov-report=html --cov-report=xml"
testpaths = ["tests"]
asyncio_mode = "auto"

